Games detail page
Update : The game detail page is now responsive in all layout (did some
rework) and almost fully functional. Game list view is also tweaked.

Bugs (Mattias, I need your help)
- In the games detail pages, the ajax to select the start letter of the
publisher, developer does not work. I did something wrong here with the
conversion of the HTML
- When going to low rez (lower than 600px wide - when the side bar goes
to bottom), the java script adds/deletes in the game detail pages for
publisher, developer, year won't work. I don't know why, everything in
high rez is ok.
- Same for the insert game on the games_main page and the quick_search
page, when going to 600px wide or less, it won't work anymore
- Mattias, after your html validation of the top bar update, it seems
that in the Picasso skin the text alignment of the top buttons of the
main menu, is not ok. The text is almost aligned to the bottom of the
button, I don't think this was the case before. Can you have a look?

Still to do
- the warning messages in the game detail pages still need to be set up
(by using $_SESSION['edit_message']  like in the games_list.php file)