CSS cleanup - Side menu code seperated using SASS
@stsilversurfer

- First of all I've updated the dir structure in the technical analysis.
Please have a look, I like to discuss this with you
- gitignore is updated --> Just like smarty, SASS created a temp folder
and even some temp files, these are ignored now using the gitignore file
- I've created a mutual PARTIAL SASS file for the menu. I've started
with the side menu. This has been transfered to the _menu.scss file in
the mutual include folder. I like to discuss this with you. I consider
this a mutual file, even though I'm using some if statements for some
skin specific stuff. Maybe this can be even made better
- All skin specific main css files are now placed in the skin folder in
main_scss --> If you use SASS, only the Scss files are edited by the
developers --> I will explain this to you