CSS cleanup
Testing the css code :
- main website : HTML validation error --> the "rel" command in the
dropdown
- Cpanel : first check on all pages in all resolutions for all skins -->
when going to phone resolution (1 column view) the left nav is always at
the top. I don't remember it being like that.  --> Needs to be fixed
- mq file of TOS --> cleanup is done by @stsilversurfer , I just wonder
which way to go. The TOS mq way, or the other skin mq way, where we have
code per section on the site.
- CPANEL TOS skin --> Responsiveness in cellphone format doesn't work
anymore. Needs to be checked