CPANEL update - db logging section
This update contains a first version of the table logging. Currently,
all changes done in the GAME section are logged. You can have a first
test with this commit. You will find more info on how to set it up at
issue #124 . so still to do :
- Logging all other sections
- making the log page responsive
- add a search functionality at the log page

This commit also fixes some bugs :
- AJAX bug removed from publisher/developer field in game detail page
- bug solved in 'add screenshots' section when there are no shots linked
to a game
- added the lightbox script properly to the boxscan upload pages
- add a link to the game detail in the music page - which was missing.