CPANEL update
I cleaned up the CSS completely as discussed in issue #146 - As you
might expect, this was a lot of work, but I removed a lot of CSS (yes I
feel stupid now ;-) ). I will be testing this more in the days to come,
but at first sight, I think I got all the files. Cleaned all the CSS and
the php files. And while doing this I solved a few bugs :
- Karma sync bug removed
- User management email function --> layout issue solved on phone
resolution
- Games series --> in add games page, the 'delete series' link was not
working --> solved