CPANEL update - menu section
A first little part of the menu section :
- added menu editor to left nav
- menu search (menu_list.php) : converted to new layouts, responsive,
html checked, logged (add new set)
- Edit set page : big part converted to new layouts, reponsive, html
checked, logged (all options on the left side), first part tested.

Also, for the menu_list page, I've enhanced the search query and the
layout a bit. I added menus done by an individual (compared to only
menus by crews) and added an indvidual column (instead of year). I don't
know if this is a good idea, but I just want to be complete. I think I
will also have to enhance the 'edit set' pages for when a set is done by
an individual, instead of a crew.