CPANEL update - doc disks
Hello all,

A first version of the doc disks is finished. I've made it in such a
way, that we can now add game and tool docs to a menu, and choose the
doc type as well. I have based this on my new table analysis, which is a
bit different than the one in the technical doc. I will update this ASAP
in the technical doc.

So what has been done more precisely :
- numerous scripts created to clean up/correct/delete menu tables (I
lost count after 6 scripts)
- complete doc disk adding done in the menu section
- everything is logged

What is next on my to do list :
- There is still a bug when adding a tool doc, the name of the tool/doc
does not appear
- write script to remove abundant category id field from table doc type
- Test test test the menu section (bugfix, skin testing and
responsiveness)
- Clean up the layout of the menu section - I'm gonna look for some
javascript to add thourough 'HELP' popups in all parts of the menu
section (I want to see if I can get a more professional look, and since
the section is so complex, some info and examples are nice)
- apart from updating the technical doc, I am thinking of adding a
complete documentation link in the CPANEL which contains parts of the
technical doc and examples
- Start with the STONISH interface
- Start with the DOC section

That will be enough for now.

Cheers,
Maarten