Fix for #201
Hello Mattias,

I checked the scripts. 2 scripts were 'problematic' as described in issue #201 : 
- nr 32  -- 2016-12-26_drop_doc_disk_game.php
- nr 20  -- 2016-12-26_drop_menu_disk_download.php
These have been removed and everything was tested using your GO_LIVE script (which is a gem!) on a vanilla prod db. The prod DB has 158 tables. There is one table - Standalone music - which is not on my dev DB. (a full documentation of all tables and a clean up of the DB is anohter little project for the future)

I also changed the falcon rgb and vga scripts, removing the autoincrement command. 

Cheers,
Maarten