Provide instructions for local installation
Added instructions on how to run the site locally and what dependencies
to install. I had to figure that stuff out to run it locally, so I
thought it may be useful to others :wink:

Also:
* Removed compiled templates from Git and exclude the full `temp/`
directory (as I understand it will be regenerated by Smarty on demand)
* Added `grunt-cli` as an NPM dependency and corresponding NPM script so
that developers don't have to have Grunt pre-installed globally (they
can just run `npm run grunt` to invoke Grunt)

This is my first pull request here, so let me know if I'm doing anything
wrong or if you disagree with it in anyway. Happy to discuss it.