Fixes #247: Fix ENTER key on the search forms
The issue was that the search forms were not true HTML `<form>`, but
rather a standalone `<input>` some Javascript with a `<a>` to trigger
the search. That prevented the ENTER key to work, as the browser will
only submit a form on ENTER if it's an actual `<form>` tag.

To fix this, converted the search to actual `<form>`s. An added
benefit is that it simplifies the code a bit, since there's no need for
Javascript to submit such forms.

Modified the CSS accordingly to account for the change of the submit
buttons from `<div>`s into proper `<input type="submit">`