Fixes #235: Interview edit screen crashing in Firefox
This was likely due to the very long SELECT for the interviewee and the
author, so replaced these with autocomplete instead.

Since we need both the "name" (to display it) and the "id" (to insert it
in the database), implemented an autocomplete endpoint that return both,
and use jQuery Autocomplete support for `label` and `value` properties.
When a row is selected by the user the label will be displayed in the
input field, and the value (id) is stored in an hidden field.

Updated the existing auto-complete endpoint to be able to auto-complete
individuals (And also added proper Content-Type header when returning
JSON data)

Updated the interview editing PHP script to also return the interview
author name (in addition to the id) since we need to display it when the
page loads