Added extra features to interview edit
Hello Nico,

Please check my addition. I'm trying to add dropdown and autocomplete, like on game main page. However, I have a hard time trying to understand what you are doing with the hidden 'members' field?? How is this filled with the newly selected author when we press the 'edit' button?

Also, the problem I have now, by adding the select code, I ofcourse have 2 member fields, one in the select, and one hidden. So when switching to the select mode, the update doesn't work. Do you have an idea how to solve this in a nice way? 

I will work some more on it tonight. (I'm doing this also to learn your code a bit of course).