[WIP] Generic auto-complete widget for dropdowns
This implements a generic autocomplete widget for the dropdowns in the
form of a jQuery plugin. It can be enabled on a select by:
* Calling `$('select').autocompleteAlternative()`
* Adding a `data-autocomplete-endpoint` attribute to the select to
specify the Ajax endpoints to use
* Adding a `data-alternative-toggle` attribute to the select to indicate
the id of an element with the button to toggle between dropdown / input mode

The auto-complete endpoint is expected to return a 2 values for each
completion, a `label` to display and the actual `value` to use in the
select (usually the ID of the entity), e.g.:

```json
{
  "label": "John Doe",
  "value": 12
}
```

The widget requires a few additional CSS classes to hide / show the
elements so I've added these too. I also added the jQuery plugin in the
main template so that it's available on all pages.