Work for #311: Setup Google Analytics tracking tags
I've used 2 different tracking tags for the main website vs. cpanel, so
that we will get separate Analytics reports.

Note that the dev server will also use the same tags so any traffic on
the dev server will be counted the same as traffic on the production
server. I don't think that's a big issue (assuming dev traffic is
insignificant), but if it becomes one we can add some code to only
output the tags on the production URL.  I'd prefer if we started simple
though and see how it goes.