Make the ENTER key work on the login form
The ENTER key didn't work on the login form because the submit button
was not a `<input type="submit">` but a `<button>`. Changed it, and
moved the password hash function on the form `onsubmit` action.

I also moved the "Remember me" checkbox just after the password rather
than after the "Sign in" button. I think that's how it's done usually,
and it has the benefit of being able to easily and quickly log in via
keyboard only:
* Type your username, then TAB to go to the password
* Type your password, then TAB to go to the "Remember me" checkbox
* Tick the checkbox with SPACE
* Hit ENTER to log in