Attempt to fix #297: "soft" 404 errors that do not set the 404 HTTP status
To fix this, the `ErrorDocument` directive in the `.htaccess` need to
point to a local file rather than an external URL.

Since the `.htaccess` files were not in the repository anymore I
re-created one and tried to rewrite the previous rules in a single file
(rather than having multiple `.htaccess` files). I also put some comments
once I deciphered what the original rules were doing.

In practice the rules do exactly the same thing has before, but the way
it's implemented is a bit different, hopefully simpler? Not sure.

Also:
* Cleaned up the error page to avoid tables, remove unnecessary
Javascript and use CSS rather than inline attributes
* Added a separate error page for 403 Forbidden, with 8 bombs this time
which means "Privilege Violation" on the TOS :wink: