Do not use `DOCUMENT_ROOT` in rewrite rules
`DOCUMENT_ROOT` is apparently not available on 1and1 hosting, and 1and1
is also weirdly configured where the document root path is not the same
as the actual path of the file as seen by PHP:
* Document root: `/kunden/homepages/xx/yyy/...`
* PHP: `/homepages/xx/yyy/...`

So reverted to the previous set of `.htaccess` files that split the
rules in 2 steps, one at the root and one under `php/`. Tweaked them a
bit so that requests to folders (i.e. `/` or `/admin/`) will work too.

Also deleted ZIP backup of the previous `.htaccess` file, now
unnecessary (and the history is in Git anyway)