Fixes #250 again: redirection on POST
Use a 307 redirect code instead of 301 so that POST data will be carried
over by the browser on redirection, which is not the case with 301.

This is especially needed on the games details page where users can
submit information about a game. The form URL is something like
`action="../../main/games/..."`, causing a rewrite to be triggered to
remove the `/main/` part.