Fixes #319: Added DB update script to convert all tables to utf8
This is a bit tricky, as the conversion needs to be done on each text
column (VARCHAR, TEXT) individually, by converting it to BLOB first,
then back to its original type.

In addition, converting a column to BLOB causes issues with some indexes
and primary keys that need be dropped then re-created afterwards.

Finally, for the `news_search_wordlist` table, because we are changing
the collation from case sensitive `latin1_bin` to case insensitive
`utf8_general_ci`, some words are not considered different anymore (e.g.
`k` and `K`), so we need to remove the duplicates and re-link the data
before re-enabling the primary key.

So I implemented a script that will:
1. Drop affected primary key constraints and indexes
2. Find all non-utf8 tables
3. For each table, find all non-utf8 columns
4. Convert the column to BLOB and back
5. Change the table default charset to utf8
6. Fixup the duplicates in `news_search_wordlist`
7. Restore the indexes and primary key constraints
8. Change the default database charset to utf8

Also, I added support for transactions in the database update system
(now that we have InnoDB tables that support transactions) to make them
safer: If any query fails during the update the whole transaction is
rolled back, so that we don't end up in an inconsistent state where some
changes were applied to the database and some weren't.

Finally once all the tables are converted, we need to make sure we
connect to the database with the right charset (utf8) by calling
`mysqli_set_charset()`. The best place to do that seemed to be
`connect.php`, but it's not in Git because it contains the database
credentials. So I split it into two: `connect.php` which just creates
the `$mysqli` instance and set the right charset, and
`connection_settings.php` which only contains the credentials. It means
that before deploying that to dev+prod, the `connection_settings.php`
file will have to be created first.