Added a rewrite rule for the old interview URLs
The Google Search Console reported that some external sites were linking
to us on an invalid URL, because they were linking interviews and the
URL to an interview has changed with the latest version of AtariLegend.

To mitigate this, add a redirection so that old URLs will continue
working, to avoid breaking links from external websites.

Example of sites linking our interviews:
* https://notes.variogr.am/page/5/
* http://www.classicarcadegaming.com/forums/index.php?topic=1089.5;wap2
* http://www.8bitrocket.com/2008/05/19/atari-history-125-atari-interviews-on-the-web/
* http://www.retrogamingtimes.com/rtm62/
* http://www.bytecellar.com/2004/10/26/time_bandit/