Move all vendor CSS files into a `vendor/` subfolder
Similar to what was done for Javascript, move the CSS dependencies into
a `vendor/` subfolder and version them.

Moved the images for LightBox next to the CSS, as it's a bit more
obvious than having them in the `includes/` folder.

I had to duplicate Font Awesome for now as I put it in each skin CSS
folder (makes more sense than in `includes/` as Font Awesome is CSS), we
can re-factor it later.