Improve page titles for SEO purposes
Ensure each page has a different titles, and try to mention something
specific in the title such as the name of the game or the name of the
person being interviewed.

That should improve our ranking and provide better search results to
user than the generic "Atari Legend | Legends Never Die".