Generate sitemaps for Google to index all of our content
Google seems to be missing some content that we offer, for example as of
yesterday it had only crawled 750 pages where it should have been a lot
more (since we have a lot more games that each have their individual
page).

To assist Google, and following their recommendations, generate
[Sitemaps](https://support.google.com/webmasters/answer/156184?hl=en)
that list all the content that need to be indexed. Where it makes sense,
also give a hint of how often the content is updated so that Google can
re-fetch it on the right schedule (rather than hammering us every day
with crawl requests).

So:
* Added a `robots.txt` to give the location of the sitemap. Also prevent
`/admin/` to be indexed, it shouldn't happen as it require
authentication but it doesn't hurt to specify it
* Generate a "sitemap index" to break down our sitemaps in multiple
files
* Generate a general sitemap containing some of the pages (e.g. News) as
well as all the interviews, and reviews page
* Generate one sitemap file per game, one for each letter of the
alphabet, 0-9 and special characters

I added a new constant `REQUEST_SITEURL` containing the actual URL used
to access the site (e.g. `localhost` on development, `dev.stonish.net`
on dev, etc.) to be able to test things: Sitemaps must contain absolute
URLs, and I couldn't test it if I hardcoded it to www.atarilegend.com.