Attributes
Made a mistake. Scratched some of the old code and added an attribute_category table to compliment the attribute_type table where attribute_category hold for example "hardware", "Display Resolution", "Tos compability" etc etc.

Need to revisit the code on the game sections in front and cpanel to take this structure into consideration.