Rename the "Class of AL" page to "About" for SEO purposes
The URL of the "Class of AL" is "classAL.php", which is not ideal for
SEO purposes as it doesn't contain any real words nor gives an idea of
what the page is about.

Rename it to just "about", as the page is about the site and Atari
Legend in general.

Note that this is really about the URL, it's not changing the page title
or the Class of AL tile which is still named "Class of AL".

Aside: Also renamed the script that animates the blocks to `about.js` to
clearly indicate it's only used in this page (I plan to use `main.js`
later for functions used across multiple pages)