Minor bug fixes and game facts update
Sorry guys, that I have done this directly into november release. These are just some minor bug fixes and an update of the game facts functions. This is not coded like I want it to be, but I just need it for the interview. Not the way I wanna work normally and I will prevent this in the future, but I want to release this cool big interview this week.