jquery-ui
Downloaded new css dependency from the jquery-ui website that is "theme-less". Then I tried to restore some of Maartens changes and placed it in a new sass file. I don't know exactly how the confirmation popup looked but Marty is a css god so he'll fix this in a jiffy.