Configure ESLint to check for Javascript issues
This configures Grunt to run ESLint to check for Javascript code issues,
both formatting and coding errors.

I had to upgrade Grunt to the latest, so make sure to run `npm install`
again on this branch.

This is integrated with the build, so if we introduce a JS issue in the
future it should show as an error on GitHub on the merge request.

For now I'm still ignoring 3 files I need to clean up (`bbcode.js`,
`forms.js` and `menus.js`) but the other have been fixed.

I used the [standard JS](https://standardjs.com/) set of rules which I
think are reasonable and not too strict, but I'm happy to discuss them
and relax some if needed.