Cleanup tabs and newlines on template files
Added a Grunt plugin to enforce the right indentation (4 spaces) and new
lines (`\n`). It's reading its config from `.editorconfig`, so it will
always be in sync with whatever changes we make to `.editorconfig`.

The painstakingly went through every HTML file to fix them :weary: .
That was a hassle, but now we should be good and with the Grunt plugin
it's guaranteed we won't re-introduce bad tabs or newlines.

Aside: Removed a few more UNUSED files