Fix path to jQuery UI images
As part of moving things around for caching purposes the `images/` directory
was moved one level too high. It should be just under the location of
the CSS file, as the CSS reference images with relative URLs like
`images/something.png`. To fix this, just move the `images/` folder
inside `vendor/` (it seems to only contain jQuery UI images).

Also renamed the jQuery UI CSS file to make it obvious that we're using
a custom theme.

Aside: Fixed a couple of `font-family` declarations to make sure the
jQuery UI buttons use the correct font from the skin.