Website tables and backend cleanup
Move the website description inside a description column: To simplify
the schema and the code, just use a `description` column on `website`
rather than having a separate `website_description` table.

Remove the unused `parent_category` column (we do not support a tree
structure for categories anymore)

Fixed the website validation system that didn't work because of a
wrong `user_id` column name.