[WIP DO NOT MERGE] DB code improvements discussion
I've been investigating how to improve our DB code, mostly to avoid very
long PHP files that are a mix of PHP logic code and SQL queries, usually
hard to follow.

I've used classes and the common [Data Access
Objects](https://en.wikipedia.org/wiki/Data_access_object) pattern, as
well as defining entity classes that maps to database tables.

I've used the links as an example, and created the following classes:
- `LinkCategory` represents a category of links, with an id, name and the
number of sites within the category. It maps to the `website_category`
table
- `LinkCategoryDAO` is the Data Access Object to access link
categories. It's in charge of querying the database to obtain the data,
and build a list of `LinkCategory` instances

In addition I've added a few DB related functions that can perform a
query with prepared statements and handle errors, so that we don't have
to constantly repeat the same `or die(...)` and the code to prepare a
statement and bind the parameters.

The main benefit is to avoid code duplication:
- Now if we need to get a list of categories anywhere else, we can just
call `LinkCategoryDAO` without having to repeat the SQL query. In that
case that will be used in the front-end to display the links with their
categories.
- If there's a bug in the SQL query it's now in a single place, easier
to fix rather than to try to locate all the PHP files using the
`website_category` table
- If we change the database (rename the tables or alter the structure)
we will only have to change the `LinkCategoryDAO` class, everything else
will mostly stay the same.

There's also a benefit in [separating
concerns](https://en.wikipedia.org/wiki/Separation_of_concerns), by
separating the SQL query from the actual coding logic in the PHP page,
so that we can focus on one or the other depending of what we're working
on. When working on the PHP page logic, we don't need to worry where the
data is coming from (as all we get is a list of `LinkCategory` objects.