Links section frontend
Implemented the frontend Links section to present all the links in the
database, by category. Also added a tile to allow logged in users to
submit a new link.

Fixed a few minor issues in the backend too, around link screenshots.

I've used a new approach for splitting the code between database access
and PHP page as discussed previously, although I haven't updated the
backend to make use of it.

Not implemented (perhaps later):
- Tracking of links clicks (to get a count / popularity of each link)
- Link rating system