Move all scripts at the bottom of the body in the frontend
Created an `additional_scripts` Smarty block that pages can use to
inject additional scripts at the end of the body. That will ensure that
all scripts used everywhere are always loaded last and don't delay the
display of the page. In turns that allows us to finally move jQuery at
the bottom of the body for faster loading.

Did the same for the CSS with `additional_css`, to remove random
`<link>` tags that got inserted in the middle of the page (where they
should be in the `<head>`).

Aside: Minor JS fixes due to confusion between the HTML comment tag
`<!--` and the JS one `//` inside `<script>` blocks.

Aside: Replaced `window.onload` with `jQuery(document).ready()` that
looks more reliable (sometimes `window.onload` didn't fire for me and
dotdotdot wouldn't apply).