Run PHP CodeSniffer by default when building
Run PHP CS on new files on the default build task, for now ignoring the
old ones (as there are too many errors to fix). Ideally everytime we add
a new PHP file we should update the Grunt config to include it in the
PHP CS task, to ensure that we don't continue to introduce syntax
violation on new code.

I kept the manual task `grunt phpcs` to check all files, not just the
new ones, in case someone feels brave enough to start fixing old files
:wink:

For now enabled it by default for `common/DAO/`, `common/Model/` and
`lib/Db.php` which are likely to be expanded as we add new DAO and Model
classes. Subsequently fixed the issues in the files existing in these
folders.

Tweaked the ruleset a bit, mostly to do with braces and a couple of
other minor things.

Updated the Docker image to install PHP CS so that Shippable can run it.