Added cache headers to image wrapper
My understanding is that images served via the image wrapper will never
change because they are tied to a specific ID in the database (so if
there's a new image, the old ID is deleted and there's a new one
generated for the new image).

That means that we can aggressively cache the image served by the image
wrapper since they will never change. Do that by adding a last modified
date and cache control header. That should help with page speed loading.