Fix a couple of issues following moving the scripts at the bottom
I missed a few things on the games list page with scripts that were
inserted in the middle of the page and relied on jQuery to be loaded in
the `<head>`. Moved them at the bottom too.

A tricky one is the "changes per month" tile because it requires jQuery
but there's no way for a tile to inject a script at the bottom of the
page. For now I worked around it by replacing `jQuery(document).ready()`
by a native JS version (Luckily, that was the only line that used
jQuery). That will be adressed better later when we refactor scripts in
separate files.

Aside: Fix the Shippable config: PHP Code Sniffer is now installed
within Docker (so that we can use it locally) so there's no need to
install it as part of the Shippable build anymore.