Create the `game_release` table
Kicking off the new structure by creating the `game_release` table and
populating it:
- Create one release per game. I left the `name` column empty since it's
supposed to be for alternative names only (e.g. a translated name for
example).
- Populate the release date from the `game_year` table. Since we are
switching from a single `year` field to a full date, I've set all dates
to the 1st of January for now.

Next step is to update the frontend and CPANEL to make use of this table
whenever the game year is displayed or modified.