Implement the `game_release` table
Kicking off the new structure by creating the `game_release` table and
populating it:
- Create one release per `game_year`. I left the `name` column empty
since it's supposed to be for alternative names only (e.g. a translated
name for example).
- Populate the release date from the `game_year` table. Since we are
switching from a single `year` field to a full date, I've set all dates
to the 1st of January for now.
- Delete the `game_year` table.

Then, updated the frontend and cpanel to make use of the `game_release`
table whenever `game_year` was used.

In a lot of cases the code was making the assumption that there was only
one `game_year` for each game, and was picking the first year: On every
table (Like the list of games, reviews, interviews), the screenstar
panel, etc.

I was unsure what to do about this, as in a lot of cases the multiple
**releases** should be listed, than just multiple dates. But it's a bit
too early to do that. So in some cases I left it as-is, in other cases I
just removed the year from the list. I think we'll need to get back to
that once the releases are a bit more fleshed out.

I updated the cpanel game editor to replace the "Release year" section
with a "Release" one, where a year can be chosen + an alternative title
for the release (which goes into `game_release.name`).

I left some sections like menus and downloads untouched, since we want
to change the DB structure for these too there was no point in fixing
them right now.