Alternative (hopefully simpler!) way to write changelog entries
Added a DAO to write to the change log to avoid having to update the
very large `create_log_entry()` function whenever we want to add an
entry to the changelog.

Most of the time when calling the function to create a changelog entry
we already have the right info for the section and sub-section. Rather
than just pass an ID, and need to write some code in
`create_log_entry()` to re-query this info, just pass it directly.

That should hopefully make it a bit less error-prone to add entries to
the changelog, especially as we'll be adding more tables in the next
months.

The DAO also "validate" the changelog entry, so that you can't add a
changelog for a section or sub-section that doesn't exist. That should
help in reminding us that we also need to update the changelog frontend
whenever we add a new section or sub-section.

I only updated the code that has to do with Users changelog entries for
now, but we should start using this version for any new table if we can.