Updating the interview section
I'm trying to optimize my code a bit (because I also like to update the reviews and articles section in the same way)
- I seperated the javascript and created a new file interviews.js
- Did some small CSS retouches (very small)
- Added the 'new' look of popups

I think the only thing I could do as rewrite the db_interviews with stored procedures, but It seems a but useless at the moment (better things to do) as everything is working nicely. Let me know if you disagree here.