Fix some of the ESLint errors on `interviews.js`
*Functions*

Functions declared outside a module are "leaked" into the global scope,
which will cause name conflict issues: If 2 scripts declare a function
named "doSomething()" for example, the second will overwrite the first
and cause bugs.

The proper fix is to use a module pattern, with a single object on the
global scope like `AtariLegend`. Then all functions would be called like
`AtariLegend.doSomething()`. Since the code would be organized in
modules inside the `AtariLegend` object, name conflicts couldn't occur.

However that's more work that we can take no now, so as a workaround
assign the function to the global `window` namespace. It doesn't solve
anything because you can still have 2 functions with the same name
assigned to `window` and it will break, but it makes ESLint happy.

*Global `previewText` function*

Same issue here, `previewText` is declared globally and leaks on the
global scope. The proper fix is to re-organize the code, but since we
can't do that work around it by telling ESLint that `previewText` is
expected to be global (which is wrong).

*Variable declarations*

Ensure all variables are declared. A variable need to be declared with a
`var` keyword, and if a variable is used inside an `if/else` block, it
should be declared beforehand.

*Syntax*

Fix a few minor syntax issue due to our code conventions (operator like
`+` must have a space after and before).

*Unused variables*

Removed unused variables.