Fix main admin game page with new DB structure
The main admin game page didn't work anymore as the DB structure has
changed.

Rather than fixing it, I've called the Ajax search when the page loads.
Since it's exactly the same query and that the Ajax search is then used
to search for games, it's better to use it on load as well rather than
having to duplicate the code in `games_main.php`.

I had to change `game_search.js` to make the search function public so
that could call it on page load.

I also fixed an issue in the Ajax HTML template that would include
`lazyload.js`. That would cause this script to be repeatedly included
again and again on every search, consuming memory and eventually
freezing the browser after 10 or 12 searches.