Refactor News code
Initially there was an issue with the Atom feed because the value
returned by `$news->getDate()` was a formatted date (i.e. "July, 2nd
2018") and not a timestamp. I started to look at `News.php` and
`NewsDAO.php` and realized there were a few issues.

Model objects like `News.php` should only contain attributes of a news
(i.e. headline, date, etc.) and nothing else, otherwise it makes them
very difficult to re-use in different pages, and makes the code more
complex and harder to maintain. The `News` model was containing
information about the news user (userid, joined date, count of news,
etc.).

Changed it to link to a `User` object instead that can hold the user
attributes. Thus accessing the id of author of a news article becomes
`$news->getUser()->getId()`. There's now `User` model that we can re-use
elsewhere, and the `News` model stays focused on news.

I think it's the first time we had something like that, so it's fair
enough that it wasn't obvious how to deal with that case.

Also:
- Remove unused / commented code
- Re-use the same template to list news on the CPANEL, regardless if
  it's done via Ajax or not (Simply by having Smarty include the Ajax
  template in the main template, rather than duplicating the code)
- Simplified the news SQL code a bit (hopefully) and added helper
  methods to deal with a variable number of constraints (i.e. constraint
  on user id + timestamp, or user id + search words, or search words +
  timestamp, etc.)
- Rather than using a separate query to get the number of results for
  the news, just use `count()` on the resulting array
- Added a DAO for users to retrieve all users
- Fixed URL of Atom feed
- Added `admin/news/*` to the PHPCS checks and cleaned up syntax warnings