Ensure `grunt watch` is always run on the Docker image
In some cases running `grunt` may fail (if PHPCS finds an error for
example) causing `grunt watch` to never be run (as it's chained with
`&&`). Replace that with `;` which will just chain the commands without
considering their exit code, so that `grunt watch` will always run even
if the initial `grunt` fails.