Fixes #567: jssor causing double page load
Not sure what went wrong, but upgrading jssor fixes the issue.

A very annoying thing is that the author of jssor claims the project is
open source under a MIT license, but doesn't make the source code
available despite a lot of people asking for it... That makes it hard to
track down issues.

Anyway, should be good now.