Fixes #576: Change game search to account for release changes
Updated game search to:
- Stop displaying the publisher under each game, since the publisher is
now on the release and a game can have multiple releases
- Similarly, stop displaying the year
- Also remove publisher and year from the export mode
- I realized that we could display the _developer_ though, because a
game still has a single developer, so I did that
- Updated the publisher search to hit the release publisher, rather than
the publisher attached to the game

Aside:
- Added a convenience button to edit a game from the main site
(when permissions allow it). I always find myself looking at a game in
the main site, wanting to edit it, and having to search it though the
CPANEL.
- Format the search time to 3 digits which is a bit nicer than the
default 11 or something.