Correct link color main site
I've removed an old class and added a new one. Because the links within text are stored with a class that I do not want to change the color codes of, I felt the look in the main site was not ok. (color was black). I did not want to change the exisiting class not to screw the look of the cpanel. So I created a new class to solve this in main site on the few spots it really bothered me, because it was hard to read.