Facebook & Twitter Meta tags
This is now already on prod and seems to be working very nicely. What a difference in facebook. Finally. 
It would be even cooler if we could auto share a picture of the game of the link we share, but for now I added a new pic which displays really nice.