Fixes #574: Make DB updates easier to work with
Change the DB update system to read instructions from individual folders
named after the DB change ID (e.g.  `database_scripts/176/my_change.ini`).

This should make the changes a bit easier to navigate as we can directly
go to a specific folder to see what the script is about, rather than
having to open all the files in turn to find the ID. There's also no
need to put the current date in the filename....

Changed the update script format to just be an INI file. It makes it a
bit simpler, safer, less verbose and less error prone, as a typo in the
PHP script could break the whole thing.

Moved all existing scripts into a `legacy/` folder. Renamed "Database
update" to "Database update (legacy)" in CPANEL and added a new
"Database update" link (which works exactly the same in practice).

Also added support for disabling foreign key checks for an update in the
INI file.

I provided 2 trivial examples of DB changes as a starting point, one
without an addition that auto-executes, and one that doesn't
auto-executes and has an addition script.