Jumpstart to get dumps for a media
Each media can have multiple dump, so we need to have an association
between a media and its dump. I chose to use an associative array for
this, where the key is the `media_id`, and the values are the Dump
objects.

In the template, just retrieve the dumps for the specific ID from the
associative array, and loop over it.

Aside: I renamed the Hash property Sha512 to be consistent with the
database name column.