For #662: Delete "unknown" files from the DEV server
So far we were only copying all files, rather than properly "syncing"
then, meaning deleting files that are on DEV, but not in Git anymore.
Fix this by passing the `--delete` flag to rsync.

Exclude `data/` and the connection settings file from deletion, as we
want to preserve these on the DEV server.