Run the DB upgrade script after syncing the files to DEV
Use the PHP command-line interface to run the script remotely (via SSH).

It's a bit messy as there are warnings because a lot of the code expects
to be run in a web environment rather than via the command-line, but it
does the job.

I had to use the PHP 7.1 CLI as 1and1 only has v5.5 or v7.1, but not 5.6
(which is the version we use on the site). v5.5 didn't work.